Avoid unwanted narrowing from long to int in tests
Some variables that receive the value of PyLong_AsLong were typed as
int which could throw away high-order bits masking the presence of an
error.

Based on Facebook D14625583

+++ ext/Objects/moduleobject-test.cpp
@@ -20,7 +20,7 @@ TEST_F(ModuleExtensionApiTest, SpamModule) {
   };
 
   // PyInit_spam
-  const int val = 5;
+  const long val = 5;
   {
     PyObject* m = PyModule_Create(&def);
     PyObject* de = PyDict_New();
@@ -37,7 +37,7 @@ TEST_F(ModuleExtensionApiTest, SpamModule) {
   PyRun_SimpleString("x = spam.CONST");
 
   PyObject* x = testing::moduleGet("__main__", "x");
-  int result = PyLong_AsLong(x);
+  long result = PyLong_AsLong(x);
   ASSERT_EQ(result, val);
 }
 
@@ -627,7 +627,7 @@ x = foo.noargs()
 )");
 
   PyObjectPtr x(moduleGet("__main__", "x"));
-  int result = PyLong_AsLong(x);
+  long result = PyLong_AsLong(x);
   ASSERT_EQ(result, 10);
   EXPECT_EQ(PyErr_Occurred(), nullptr);
 }
@@ -669,7 +669,7 @@ x = foo.varargs(10)
 )");
 
   PyObjectPtr x(moduleGet("__main__", "x"));
-  int result = PyLong_AsLong(x);
+  long result = PyLong_AsLong(x);
   ASSERT_EQ(result, 10);
   EXPECT_EQ(PyErr_Occurred(), nullptr);
 }
@@ -700,7 +700,7 @@ x = foo.kwArgs(value=40)
 )");
 
   PyObjectPtr x(moduleGet("__main__", "x"));
-  int result = PyLong_AsLong(x);
+  long result = PyLong_AsLong(x);
   ASSERT_EQ(result, 40);
   EXPECT_EQ(PyErr_Occurred(), nullptr);
 }
